Update method signatures
